Fix install to source bash files and only include bash functions from project directory (no need for symlink)

+++ install.sh
@@ -27,10 +27,12 @@ mergeddir=$projectdir/$mergeddir;
 
 installfile()
 {
+    echo " - == Installing $1";
+
     # Get filenames
     localname=$projectdir/$1;
     backupname=$backupdir/$1;
-    mergedname=$mergeddir/$1;
+    source=$3;
 
     # If a different option was supplied for install name, use that
     if [ ! -z $2 ]; then
@@ -39,39 +41,46 @@ installfile()
         installname=$installdir/$1;
     fi
 
-    # Backup existing file
-    if [ -e $installname ]; then
-        # If it was our simlink, just remove it
-        link=`readlink -f $installname`;
-        if [ "${link#*$projectdir}" != "$link" ]; then
-            rm $installname && echo " - Removed old link to .shellconfig";
-        fi
+    link=`readlink -f $installname`;
 
+    # Doesn't exist, create symlink (simplest)
+    if ! [ -e $installname ]; then
+        ln -s $localname $installname && echo " - Created symlink to .shellconfig";
+
+    # If one of our links
+    elif [ -L $installname ] && [ "${link#*$projectdir}" != "$link" ]; then 
+        # Remove old link and create new one
+        rm $installname && echo " - Removed old link to .shellconfig";
+	ln -s $localname $installname && echo " - Created symlink to merged file";
+
+    # If we're to source it
+    elif [ -e $installname ] && [ $source ]; then
+        echo "source $localname; #nottrobin-shellconfig" >> $installname && echo " - Sourced from .shellconfig";
+
+    # Exists but not sourceable
+    elif [ -e $installname ]; then
         # Make directories if they don't exist
         [ ! -e $backupdir ] && mkdir $backupdir && echo " - Made backup directory";
-        [ ! -e $mergeddir ] && mkdir $mergeddir && echo " - Made merged directory";
-        
+
         # Back up the existing file
-        cp $installname $backupname && echo " - Backed up original file";
+        cp -r $installname $backupname && echo " - Backed up original file";
 
-        # Create merged file
-        cat $installname $localname > $mergedname && echo " - Created merged file";
-        
         # Remove original file
-        rm $installname && echo " - Removed original file";
+        rm -r $installname && echo " - Removed original file";
 
-	# Link to the merged version
-	ln -s $mergedname $installname && echo " - Created symlink to merged file";
-    else
 	# Link to our version
-	ln -s $localname $installname && echo " - Created symlink to .shellconfig";
+	ln -s $localname $installname && echo " - Created symlink to our file";
     fi
-
 }
 
 # Setup normal files
-for filename in '.bashrc' '.bash_profile' '.bash_functions.sh' '.tmux.conf' '.vim'; do
+for filename in '.tmux.conf' '.vim'; do
     installfile $filename && echo " - == Installed $filename";
+done
+
+# Source bash files
+for filename in '.bashrc' '.bash_profile'; do
+    installfile $filename $filename true && echo " - == Installed $filename";
 done
 
 # Setup global git config (needs special name)
